tag command for Go
Add 'releasetool tag' for Go.

This command is quite different for Go, because GitHub is not the
source of truth, and also because a "release" for Go is just tagging
the repo.

Rather than mess with the Gerritt API, we work off the local repo, and
push the tag at the end.