feat: Improve .NET tag command handling (#249)
* feat: Improve PR message matching for .NET

- We now match the title of the PR as well as the body
- The "- " is optional (mostly for the sake of matching the conventional PR title)
- We match at the start of the line, so "Don't Release Foo version Bar" won't match

* feat: Improve the .NET GitHub release format

- Our releases now have titles of just "Foo version Bar"
- Copy the body of the PR as the body of the release

Note that the current .NET TagReleases code performs reformatting
from the commit message, so that multiple lines are "unwrapped" into
single long lines, etc. We can move that functionality into
ReleaseManager in a new "push" command that will create the PR (with
an autorelease label), so the PR body should already be fine for the
release.