fix: update kokoro release job name for google-cloudevents-ruby
Switching the kokoro release job for google-cloudevents-ruby from one located in the `cloud-devrel/ruby` directory to one located `cloud-devrel/client-libraries` directory. (The kokoro configs have already been duplicated into the new location.) This move is an attempt to provide the correct GitHub credentials for the release job to update the release pull request: The client-libraries directory has access to resources for repositories under the googleapis org, resources that seem not to be available to other subdirectories of cloud-devrel.